Finish Company Dashboard

+++ src/main/java/acme/forms/CompanyDashboard.java
@@ -13,7 +13,7 @@ public class CompanyDashboard extends AbstractForm {
 
 	private static final long		serialVersionUID	= 1L;
 
-	protected Map<String, Integer>	nPracticumByMonthLastYear;
+	protected Map<String, Long>		nPracticumByMonthLastYear;
 	protected Map<String, Double>	statisticsPeriodLengthOfTheSessionsInTheirPractica;//average,deviation,min,max
 	protected Map<String, Double>	statisticsPeriodLengthOfTheirPractica;
